api/wwcodesvtools/requirements.txt
Django==3.1
psycopg2==2.8.5
South==1.0.2
django-registration==3.1
gunicorn==20.0.4
django-environ==0.4.5
djangorestframework==3.11.1
pre-commit==2.7.1
djangorestframework-simplejwt==4.6.0
drf-yasg==1.17.1
django-cors-headers==3.5.0
django_heroku==0.3.1